внесены исправления по ошибкам, логика movies и saved movies разделена
