+++ src/components/Register/Register.js
@@ -3,6 +3,7 @@ import './Register.css';
 import { useFormAndValidation } from '../../hooks/useFormAndValidation';
 import LoginRegisterForm from '../LoginRegisterForm/LoginRegisterForm';
 import { useNavigate } from "react-router-dom";
+import { EMAIL_PATTERN } from "../../utils/constants";
 
 const Register = ({ isLoggedIn, handleRegister }) => {
 
@@ -60,6 +61,7 @@ const Register = ({ isLoggedIn, handleRegister }) => {
                 id='email'
                 className={`form__input ${errors.email && 'form__input_color_red'}`}
                 value={values.email || ''}
+                pattern={EMAIL_PATTERN}
                 required
             />
             {!isValid && errors.email && <span className="form__label form__label_type_error">{`${errors.email}`}</span>}
